+++ backend/src/Middleware/Guzzle/Esi429Response.php
@@ -42,6 +42,7 @@ class Esi429Response
 
     private function handleResponseHeaders(ResponseInterface $response): void
     {
+        // See also https://github.com/esi/esi-issues/issues/1227
         if ($response->getStatusCode() === 500) {
             $body = $response->getBody()->__toString();
             if (
@@ -51,9 +52,29 @@ class Esi429Response
                 strpos($body, 'You have been temporarily throttled.') !== false
             ) {
                 $this->storage->set(Variables::ESI_THROTTLED, '1');
-                return;
             }
+        } else {
+            $this->storage->set(Variables::ESI_THROTTLED, '0');
         }
-        $this->storage->set(Variables::ESI_THROTTLED, '0');
+
+        // SSO rate limit, see also https://developers.eveonline.com/blog/article/sso-endpoint-deprecations-2
+        if ($response->getStatusCode() === 429) {
+            $waitUntil = time() + 60;
+            if ($response->hasHeader('Retry-After')) {
+                $retryAfter = $response->getHeader('Retry-After')[0];
+                $this->logger->warning("Esi429Response Retry-After: $retryAfter");
+                if (is_numeric($retryAfter)) {
+                    // e.g.: 120
+                    $waitUntil = time() + $retryAfter;
+                } else {
+                    // e.g.: Wed, 21 Oct 2015 07:28:00 GMT
+                    $datetime = \DateTime::createFromFormat('D, d M Y H:i:s T', $retryAfter);
+                    if ($datetime instanceof \DateTime) {
+                        $waitUntil = $datetime->getTimestamp();
+                    }
+                }
+            }
+            $this->storage->set(Variables::ESI_RATE_LIMIT, (string) $waitUntil);
+        }
     }
 }